requirements.txt
matplotlib
pandas
azure-cognitiveservices-speech
python-dotenv
openai
